Split out the apt repo key download
The apt_key module is not really properly equipped for the key
downloads because it can only handle the happy path scenarios. In
order to get full control over the download, we instead opted to use
the uri module to perform the download. This will allow us to expand
the download part in the future when we add support for
packagercloud's read tokens.