Add support for packagecloud's read tokens
Up until now, it was not possible to install Sensu Go from a private
repositories on packagecloud. To overcome that deficiency, we added
another variable to the install role that should contain read token.

Note that we currently do not have any tests for this functionality
because we do not have access to any of the private repositories.
Hopefully, this will get fixed in the future.