Delegate requirements installation ansible-test tool
Up until now, we manually installed all test requirements. Why?
Because historically, ansible-test was installing a buch of stuff that
was not relevant to our tests when asked to manage its requirements.

But things changed for the better in last year and now we are ready to
drop most of the custom requirements and delegate the management to
the ansible-test.

Note that there are still some packages that we need to install
manually because our tests go a bit further than what ansible-test
offers. Those are still present in requirement files. But we simulate
the ansible-test even here and install them on-demand.