Deprecate custom entity classes
Sensu Go officially supports two entity classes: agent and proxy.
Using any other class can lead to undefined behavior of the backend
and should not be used in production.

In order to give Ansible users a heads-up about this issue, we added a
deprecation warning into Sensu Go Ansible Collection wen unsupported
entity class is passed in as a parameter.