Update Molecule
Molecule 3.1.0 was the first one that supports ansible-base. This is
the first step towards getting our tests running on ansible-base.

Another reason for updating Molecule is our planed transition to
pytest-molecule as a test runner. This will allow us to automatically
time our test executions, allowing us to get rid of the custom
scheduling logic that we currently have.