Use pytest for integration test discovery
Up until now, we invoked Molecule directly when running integration
tests. And it is still useful to run it manually when developing a
module or a role since we often need to interactively log into the
target container.

But when running tests in bulk, pytest brings quite a few advantages.
One is the automatic test discovery,which means that we no longer rely
on makefile wildcards to list all of our tests.

But even more importantly, pytest allows us to automatically time the
scenario times, thich makes is a great fir with the CircleCI's ability
to split the tests across multiple workers automatically based on
timings from the previous runs. This way, we can get rid of all our
custom partitioning logic that we used before and manual timing of
tests.