Split default installation scenario
We use default scenario primarily to test if our repository
configuration works. And because we need to test this of a lot of
different distributions and versions, the scenario has a lot of
platform entries.

In order to make it a bit easier and faster to run tests, we split the
default installation scenario into two parts: one for RPM-based
systems, and the other one for the DEB-based distros.

At the moment, the split is a bit uneven, but things will balance a
bit better soonish.