Add support for installing agent on Windows
From now on, we can use the installation role for installing Sensu Go
agents on Windows hosts.

The versions we support are hardcoded in the collection. We had to do
this because we need to know product code in advance. This means that
we will need to release Sensu Go Ansible Collection a bit more often
in order to keep up with the upstream developers.

Do note that adding new version is not too problematic. What we need
to do is:

 1. Find the version and build we are interested in on [1].
 2. Download the x64 and x86 msi files from the download page (use the
    download URL from [2] as a base and adjust as needed).
 3. Extract product codes from the msi packages using some kind of msi
    editor.
 4. Add new entry to the roles/install/vars/Windows.yml file.

[1] https://packagecloud.io/sensu/stable
[2] https://docs.sensu.io/sensu-go/latest/operations/deploy-sensu/install-sensu/#install-sensu-agents