Add support for hashed passwords in user module
Sensu Go's API allows us to operate using hashed passwords in certain
situations. In order to align with the upstream's capabilities, we
added password_hash argument to the user module.

Do note that change detection does not work when using hashed
passwords because the backend does not return back enough information
for us to determine if anything changed.