Prepare user module for unversioned backends
When the Sensu Go backend is compiled with default settings, no
version number is added to the source. Previously, this bad version
number caused user module to fail because default version value
(devel) is not a valid version number.

Changes in this commit handle such version numbers as something waaaay
in the future. This allows us to keep all of our logic intact since we
are dealing with a valid version. It just happens to be a very large
one ;)