Move unit tests
We initially misplaced our units tests (we missed the plugins part of
the path). And because ansible-test is using paths to determine which
set of constraints certain tests must fulfil, we were missing out on
some additional safety.

But the good news is, all of the tests still pass even wit more strict
requirements.