Update testing tools
This update allows us to run our tests against a development version
of Ansible.

Do note that some tools we use did not have a proper release yet and
are still under development. We will replace them with a stable
version of the tool as soon as the changes are released.