Add ansible.windows dependency
When we added support for Sensu Go agent installation on Windows, we
forgot to include the ansible.windows dependency.

Why our CI/CD system did not catch this? Because our previous versions
of testing tools had dependency on the ansible package that provides
Windows modules, the missing dependency was hidden from us. Now that
we started using ansible-base and ansible-core for testing, the lack
of dependency became obvious.