Make sure package downgrades work as expected
In order to make sure package downgrades work as intented, we added an
integration test that performs a downgrade process.

Once we run the test, we discovered that apt and yum are having
troubles with the downgrade process. Fixing yum issue was easy: we
just had to replace the `state: latest` with the `state: present` when
we had a version number specified.

Fixing the apt issue was a bit more tricky because the apt Ansible
module does not have a parameter for allowing downgrades. For the time
being, we resolved this by adding a `force: true` to the apt
invocation. We should remove this temporary hack as soon as the
upstream adds options for downgrading packages.