Lazy-load all Windows-related content
When we released 1.9.3 and run integration tests that simulate user
installing the Sensu Go Ansible Collection from galaxy, we found out
that we messed up. Because the Windows Ansible Collection in not a
dependency anymore, Ansible refused to parse our playbook because it
could not find a win_service module.

Most of the Windows-related functionality is already lazy-loaded (we
include it using include_tasks directive), but there was one place
that we eagerly loaded win_service module: Sensu Go agent service
handler.

This commit replaces the eager variant with a lazy alternative that
will only fail if the handler is actually executed. This simulates how
include_tasks directive works and thus does not introduce
inconsistencies into our collection.