Simulate package installation in integration tests
It has happened twice now that we only detected some packaging
breakage during the stable branch testing because the mater branch was
tested in a different way from the regular usage.

This is why we started to test things against an installed Ansible
collection. This means that our development version of collection is
first packaged and installed before we start executing integration
tests.