Add ansible-core-2.11 to testing matrix
Ansible branched the next stable version from the development branch
yesterday. This means that the devel branch now contains code that
will become ansible-core-2.12 while the sources for ansible-core-2.11
now live in the stable-2.11 branch.

To keep our CI/CD working, we added stable-2.11 to the test matrix.
And to keep the development version test clean, we copied over the
ignore file.