Add support for remote bonsai access
Up until now, we retrieved asset definitions from Sensu Bonsai on the
control node. The reasoning for this decision was that here is a
higher probability that the control node has access to the
bonsai.sensu.io.

This commit allows users to select where the asset definitions should
be downloaded. By default, we still download them on control node, but
for cases where the target node is more capable, we allow users to
specify that they want to run the download process on target node.

How does it work? The bonsai_asset action plugin now checks the
on_remote parameter and if the parameter is missing or is set to
false, performs exactly the same thing as it did before (directly
download the asset definition).

If the on_remote parameter is set to true, the bonsai_asset action
plugin will execute a helper module. The only job of that module is to
retrieve the data on the target node and send back the asset
definition to control node. This helper module is an implementation
detail and is not exposed to the users. This is why it is OK to have a
mismatch in the actual module implementation and the documentaton.